feat(ErrorMessage): ErrorMessage Enum 정의
