+++ src/main/java/lotto/message/ErrorMessage.java
@@ -1,0 +1,16 @@
+package lotto.message;
+
+public enum ErrorMessage {
+
+    PREFIX("[ERROR] ");
+
+    private final String message;
+
+    ErrorMessage(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
